[MIG] Version 12.0
