+++ pos_order_picking_link/README.rst
@@ -1,44 +1,98 @@
-.. image:: https://img.shields.io/badge/licence-AGPL--3-blue.svg
-   :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
-   :alt: License: AGPL-3
+======================
+POS Order Picking Link
+======================
 
-=================================================
-Link between the picking and its source POS order
-=================================================
+.. !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+   !! This file is generated by oca-gen-addon-readme !!
+   !! changes will be overwritten.                   !!
+   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
-This module adds a link between pickings and its source POS order.
-So that user can easily reach the pos order related to the picking
-and see the pickings related to the pos order. This only applies
-on the pickings generated from pos order.
+.. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
+    :target: https://odoo-community.org/page/development-status
+    :alt: Beta
+.. |badge2| image:: https://img.shields.io/badge/licence-LGPL--3-blue.png
+    :target: http://www.gnu.org/licenses/lgpl-3.0-standalone.html
+    :alt: License: LGPL-3
+.. |badge3| image:: https://img.shields.io/badge/github-OCA%2Fpos-lightgray.png?logo=github
+    :target: https://github.com/OCA/pos/tree/12.0/pos_order_picking_link
+    :alt: OCA/pos
+.. |badge4| image:: https://img.shields.io/badge/weblate-Translate%20me-F47D42.png
+    :target: https://translation.odoo-community.org/projects/pos-12-0/pos-12-0-pos_order_picking_link
+    :alt: Translate me on Weblate
+.. |badge5| image:: https://img.shields.io/badge/runbot-Try%20me-875A7B.png
+    :target: https://runbot.odoo-community.org/runbot/184/12.0
+    :alt: Try me on Runbot
 
+|badge1| |badge2| |badge3| |badge4| |badge5| 
+
+Adds link between Pos orders and generated pickings
+
+**Table of contents**
+
+.. contents::
+   :local:
+
+Usage
+=====
+
+Go to a picking and see the linked Pos Order
 
 Bug Tracker
 ===========
 
 Bugs are tracked on `GitHub Issues <https://github.com/OCA/pos/issues>`_.
 In case of trouble, please check there if your issue has already been reported.
-If you spotted it first, help us smashing it by providing a detailed and welcomed feedback
+If you spotted it first, help us smashing it by providing a detailed and welcomed
+`feedback <https://github.com/OCA/pos/issues/new?body=module:%20pos_order_picking_link%0Aversion:%2012.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+
+Do not contact contributors directly about support or help with technical issues.
 
 Credits
 =======
 
+Authors
+~~~~~~~
+
+* KMEE
+
 Contributors
-------------
+~~~~~~~~~~~~
 
-* Daniel Sadamo <daniel.sadamo@kmee.com.br>
+* `KMEE <https://kmee.com.br>`__:
+    * Luis Felipe Mileo <mileo@kmee.com.br>
+    * Daniel Sadamo <daniel.sadamo@kmee.com.br>
 
+Other credits
+~~~~~~~~~~~~~
 
-Maintainer
-----------
+The development of this module has been financially supported by:
 
-.. image:: http://odoo-community.org/logo.png
-   :alt: Odoo Community Association
-   :target: http://odoo-community.org
+* `KMEE <https://kmee.com.br>`__:
+
+Maintainers
+~~~~~~~~~~~
 
 This module is maintained by the OCA.
+
+.. image:: https://odoo-community.org/logo.png
+   :alt: Odoo Community Association
+   :target: https://odoo-community.org
 
 OCA, or the Odoo Community Association, is a nonprofit organization whose
 mission is to support the collaborative development of Odoo features and
 promote its widespread use.
 
-To contribute to this module, please visit http://odoo-community.org.
+.. |maintainer-sadamo| image:: https://github.com/sadamo.png?size=40px
+    :target: https://github.com/sadamo
+    :alt: sadamo
+.. |maintainer-mileo| image:: https://github.com/mileo.png?size=40px
+    :target: https://github.com/mileo
+    :alt: mileo
+
+Current `maintainers <https://odoo-community.org/page/maintainer-role>`__:
+
+|maintainer-sadamo| |maintainer-mileo| 
+
+This module is part of the `OCA/pos <https://github.com/OCA/pos/tree/12.0/pos_order_picking_link>`_ project on GitHub.
+
+You are welcome to contribute. To learn how please visit https://odoo-community.org/page/Contribute.
