+++ pos_order_picking_link/__init__.py
@@ -1,5 +1,4 @@
-# -*- coding: utf-8 -*-
 # © 2016 KMEE INFORMATICA LTDA (https://kmee.com.br)
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+# License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl.html).
 
 from . import models
