+++ pos_order_picking_link/__manifest__.py
@@ -1,0 +1,17 @@
+# © 2016 KMEE INFORMATICA LTDA (https://kmee.com.br)
+# License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl.html).
+
+{
+    'name': 'POS Order Picking Link',
+    'version': '12.0.1.0.0',
+    'category': 'Point of Sale',
+    'author': 'KMEE, Odoo Community Association (OCA)',
+    'website': 'https://github.com/oca/pos',
+    'license': 'LGPL-3',
+    'depends': ['point_of_sale'],
+    'data': [
+        'views/stock_view.xml',
+    ],
+    'maintainers': ['sadamo', 'mileo'],
+    'installable': True,
+}
