+++ pos_order_picking_link/readme/CONTRIBUTORS.rst
@@ -1,0 +1,3 @@
+* `KMEE <https://kmee.com.br>`__:
+    * Luis Felipe Mileo <mileo@kmee.com.br>
+    * Daniel Sadamo <daniel.sadamo@kmee.com.br>
