+++ pos_order_picking_link/readme/CREDITS.rst
@@ -1,0 +1,3 @@
+The development of this module has been financially supported by:
+
+* `KMEE <https://kmee.com.br>`__:
