+++ pos_order_picking_link/readme/DESCRIPTION.rst
@@ -1,0 +1,1 @@
+Adds link between Pos orders and generated pickings
