+++ pos_order_picking_link/readme/USAGE.rst
@@ -1,0 +1,1 @@
+Go to a picking and see the linked Pos Order
