+++ pos_order_picking_link/views/stock_view.xml
@@ -1,18 +1,18 @@
 <?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data>
+<odoo>
 
-        <!-- stock.picking -->
-        <record id="view_picking_form" model="ir.ui.view">
-            <field name="name">stock_picking_invoice_link.stock.picking.form</field>
-            <field name="model">stock.picking</field>
-            <field name="inherit_id" ref="stock_account.view_picking_inherit_form2"/>
-            <field name="arch" type="xml">
-                <field name="invoice_state"  position="after">
-                    <field name="pos_order_ids" widget="many2one"/>
-                </field>
-            </field>
-        </record>
+    <record id="view_picking_form" model="ir.ui.view">
+        <field name="name">stock_picking_invoice_link.stock.picking.form</field>
+        <field name="model">stock.picking</field>
+        <field name="inherit_id" ref="stock.view_picking_form"/>
+        <field name="arch" type="xml">
+            <notebook position="inside">
+                <page name="pos_orders" string="Point of Sale">
+                    <field name="pos_order_ids"/>
+                </page>
+            </notebook>
+        </field>
+    </record>
 
-    </data>
-</openerp>
+
+</odoo>
